Update to backend lookup tables (npz)
This was a big update to the backend of how redshifts are calculated.
Now the interpolation is performed at the time of requesting redshift
rather than to create pre-calculated tables.

* The lookup tables are now stored as npz files instead of saved
instance methods of interp1d. The reason for this change is three fold:
1. The size of the npz is less than half that of npy.
2. This show allow python 2 users to use table.create
3. It should be easier to create functions that will create a redshift
DM relations if all that is required is an output array rather than an
explicit function.
* Added lookup_table as a parameter to calc_redshift()
* Added create_methods_catalogue() to catalogue
* Added get_z_from_table() to table
* Added tests for catalogue and comoving_distance()